requirements.txt
Flask==2.2.2
google-generativeai==1.0.0
pyttsx3==2.90
SpeechRecognition==3.8.1
pdfplumber==0.6.1
torch
torchvision
easyocr
pyaudio
pymongo
dotenv